SIS is now a double, and is printed with 2 decimals

+++ tree.h
@@ -62,7 +62,7 @@ typedef struct __lih_SDIinfo
 	TreePtr *lost; /**< array of lost genes */
 	int n_spec_node; /**< used for tr_SDI() */
 	int is_confirmed_dup;
-	int sis; /* species intersection score */
+	double sis; /* species intersection score */
 	char *spec_node; /**< used for tr_SDI() */
 } SDIinfo;
